so, get everything back to a working state:
1. make the code re: generic types no longer stateful. change the name of generic types before generation so that the "most-recently-generated-type" can't be looked up from the basename (relax, no mangledName-based-lookup)
2. propagate the necessity of the generic type map to more places in the code; basically, if you know about generic types, then you are required to pass on that knowledge to stuff that you touch (if they need it)
3. enabled warnings on MSVC, and fix a couple of those signedness things.

+++ source/codegen/glue/strings.cpp
@@ -895,8 +895,8 @@ namespace string
 				// if statement
 				auto ch = cs->irb.Load(cs->irb.Load(ptrp));
 
-				auto mask = cs->irb.BitwiseAND(ch, fir::ConstantInt::getInt8(0xC0));
-				auto isch = cs->irb.ICmpNEQ(mask, fir::ConstantInt::getInt8(0x80));
+				auto mask = cs->irb.BitwiseAND(ch, fir::ConstantInt::getInt8((int8_t) 0xC0));
+				auto isch = cs->irb.ICmpNEQ(mask, fir::ConstantInt::getInt8((int8_t) 0x80));
 
 				fir::IRBlock* incr = cs->irb.addNewBlockInFunction("incr", func);
 				fir::IRBlock* skip = cs->irb.addNewBlockInFunction("skip", func);
